Remove duplicate registrar modules
