--- pkg/mediorum/httputil/httputil.go
@@ -1,11 +1,0 @@
-package httputil
-
-import "regexp"
-
-var (
-	trailingSlash = regexp.MustCompile("/$")
-)
-
-func RemoveTrailingSlash(s string) string {
-	return trailingSlash.ReplaceAllString(s, "")
-}
