--- pkg/mediorum/registrar/the_graph_test.go
@@ -1,23 +1,0 @@
-package registrar
-
-import (
-	"fmt"
-	"testing"
-)
-
-func TestTheGraph(t *testing.T) {
-	t.Skip()
-
-	{
-		p := NewGraphProd()
-		fmt.Println(p.Peers())
-		fmt.Println(p.Signers())
-	}
-
-	{
-		p := NewGraphStaging()
-		fmt.Println(p.Peers())
-		fmt.Println(p.Signers())
-	}
-
-}
