+++ pkg/registrar/multi.go
@@ -2,6 +2,9 @@ package registrar
 
 import (
 	"errors"
+	"fmt"
+
+	"github.com/AudiusProject/audius-protocol/pkg/mediorum/server"
 )
 
 func NewMultiStaging() PeerProvider {
@@ -9,6 +12,7 @@ func NewMultiStaging() PeerProvider {
 		providers: []PeerProvider{
 			NewAudiusApiGatewayStaging(),
 			NewGraphStaging(),
+			NewEthChainProvider(),
 		},
 	}
 
@@ -19,6 +23,7 @@ func NewMultiProd() PeerProvider {
 		providers: []PeerProvider{
 			NewAudiusApiGatewayProd(),
 			NewGraphProd(),
+			NewEthChainProvider(),
 		},
 	}
 }
@@ -27,14 +32,23 @@ type multiProvider struct {
 	providers []PeerProvider
 }
 
-func (p multiProvider) Peers(nodeType string) ([]Peer, error) {
-	if nodeType != "discovery" && nodeType != "content" && nodeType != "all" {
-		return nil, errors.New("invalid node type")
-	}
-
+func (p multiProvider) Peers() ([]server.Peer, error) {
 	for _, provider := range p.providers {
-		if vals, err := provider.Peers(nodeType); err == nil {
+		if vals, err := provider.Peers(); err == nil {
 			return vals, nil
+		} else {
+			fmt.Println(err)
+		}
+	}
+	return nil, errors.New("all providers failed")
+}
+
+func (p multiProvider) Signers() ([]server.Peer, error) {
+	for _, provider := range p.providers {
+		if vals, err := provider.Signers(); err == nil {
+			return vals, nil
+		} else {
+			fmt.Println(err)
 		}
 	}
 	return nil, errors.New("all providers failed")
