+++ pkg/registrar/static.go
